Attempt to make pure python branch merge.
Reverted most changes in algos3d and humanmodifier since rev 3593, except for those necessary to make warping work. The human is accessed through a global variable, which is not so pretty but works. The isWarp attribute is accessed through a hasattr. Hopefully the changes to human.py will be acceptable.