Warping update.
Warping can now be turn on and off in the expression tuning tab. This is a global switch which affects both expressions and poses. By default warping is turned off. In this case the benchmark pose looks good, so the problems evidently had to do with warping. In the release the warp method switch should be hidden, but for the time being this will allow me to figure out what goes wrong without affecting other people.

Warped targets are no longer written to a temporary file.