The DAZ preset has been removed from makewalk for copyright reasons.
However, bvh files targeting DAZ can still be loaded, because a new detection system for source armatures has been introduced. The new system is based on armature topology rather than bone names. It will only work if the rest pose has head up and face forward. The new and old systems coexist for the time being. If no matching .src file is found, the new system is used instead.

The new system only works if the bvh armature's rest pose is T-pose. This restriction should be lifted in the future, but the DAZ case is still covered.