Some tweaks of the proxy handling code. The available simple proxy types are now defined by a list at the top of mh2proxy. The types are capitalized, by a lowercase version is also defined since the material code uses that. Perhaps that should be changed in the future.
All combinations of proxy types, materials and exporters seem to work, except that the proxymesh material has no texture on export. I have not tested the renderers, which seem to need a major overhaul anyway. The code is aware of eyelashes, teeth and tongue as well, so it should work out of the box as soon as they become available.