MHX rig: All properties and drivers related to file linking have been completely removed. Shapekeys and visibility are thus controlled in the standard way on the mesh.
A character can still be set up for file linking, but this is now done on the Blender side by a new tool called sandbox, which is located in utils/tools. As the name indicates, it contains various experimental and undocumented tools, which may be made official some day. For the time being, the tool should be unobtrusive, since it is located under utils and thus not included in releases and nightly builds.